backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
google-cloud-aiplatform==1.36.4
langchain==0.0.335
langchain-google-genai==0.0.5
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
python-docx==1.0.1
PyPDF2==3.0.1
openpyxl==3.1.2
pillow==10.1.0
pytesseract==0.3.10 